= 1.0.5 - 10 March, 2021 =
- Tweak: Changed perfectscroll library to conditional.
- Tweak: Quick view modal design.
- Tested: with WordPress version 5.9.2.
- Updated: Dev tools (NPM & GULP).
