+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "addonify-quick-view",
-  "version": "1.0.0",
-  "description": "Addonify WooCoomerce Quick View plugin adds functionality to have a WooCoomerce product quick view preview on a modal window.",
-  "main": "index.js",
+  "version": "1.0.4",
+  "description": "Addonify WooCommerce Quick View plugin adds functionality to have a WooCommerce product quick preview on a modal window.",
+  "main": "gulpfile.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -10,23 +10,29 @@
     "type": "git",
     "url": "git+https://github.com/addonify/addonify-quick-view.git"
   },
-  "keywords": [
-    "woocommerce",
-    "quick",
-    "view",
-    "addonify"
-  ],
-  "author": "Addonify",
+  "author": "addonify",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/addonify/addonify-quick-view/issues"
   },
   "homepage": "https://github.com/addonify/addonify-quick-view#readme",
   "devDependencies": {
+    "autoprefixer": "^10.4.2",
+    "cssnano": "^5.1.1",
     "gulp": "^4.0.2",
-    "gulp-wp-pot": "^2.4.3"
-  },
-  "dependencies": {
-    "node-modules": "^1.0.1"
+    "gulp-concat": "^2.6.1",
+    "gulp-notify": "^4.0.0",
+    "gulp-plumber": "^1.2.1",
+    "gulp-postcss": "^9.0.1",
+    "gulp-rename": "^2.0.0",
+    "gulp-replace": "^1.1.3",
+    "gulp-rtlcss": "^1.4.2",
+    "gulp-sass": "^5.1.0",
+    "gulp-shell": "^0.8.0",
+    "gulp-sourcemaps": "^3.0.0",
+    "gulp-uglify": "^3.0.2",
+    "gulp-wp-pot": "^2.5.0",
+    "gulp-zip": "^5.1.0",
+    "sass": "^1.49.9"
   }
 }
