+++ public/assets/src/scss/conditional/perfect-scrollbar.scss
@@ -1,5 +1,6 @@
 
 .ps {
+
     overflow: hidden !important;
     overflow-anchor: none;
     -ms-overflow-style: none;
@@ -11,6 +12,7 @@
  * Scrollbar rail styles
  */
 .ps__rail-x {
+
     display: none;
     opacity: 0;
     transition: background-color .2s linear, opacity .2s linear;
@@ -23,6 +25,7 @@
 }
 
 .ps__rail-y {
+
     display: none;
     opacity: 0;
     transition: background-color .2s linear, opacity .2s linear;
@@ -36,6 +39,7 @@
 
 .ps--active-x>.ps__rail-x,
 .ps--active-y>.ps__rail-y {
+
     display: block;
     background-color: transparent;
 }
@@ -46,6 +50,7 @@
 .ps--focus>.ps__rail-y,
 .ps--scrolling-x>.ps__rail-x,
 .ps--scrolling-y>.ps__rail-y {
+
     opacity: 0.6;
 }
 
@@ -55,6 +60,7 @@
 .ps .ps__rail-y:focus,
 .ps .ps__rail-x.ps--clicking,
 .ps .ps__rail-y.ps--clicking {
+
     background-color: #eee;
     opacity: 0.9;
 }
@@ -63,6 +69,7 @@
  * Scrollbar thumb styles
  */
 .ps__thumb-x {
+
     background-color: #aaa;
     border-radius: 6px;
     transition: background-color .2s linear, height .2s ease-in-out;
@@ -75,6 +82,7 @@
 }
 
 .ps__thumb-y {
+
     background-color: #aaa;
     border-radius: 6px;
     transition: background-color .2s linear, width .2s ease-in-out;
@@ -89,6 +97,7 @@
 .ps__rail-x:hover>.ps__thumb-x,
 .ps__rail-x:focus>.ps__thumb-x,
 .ps__rail-x.ps--clicking .ps__thumb-x {
+
     background-color: #999;
     height: 11px;
 }
@@ -96,20 +105,25 @@
 .ps__rail-y:hover>.ps__thumb-y,
 .ps__rail-y:focus>.ps__thumb-y,
 .ps__rail-y.ps--clicking .ps__thumb-y {
+
     background-color: #999;
     width: 11px;
 }
 
 /* MS supports */
 @supports (-ms-overflow-style: none) {
+
     .ps {
+
         overflow: auto !important;
     }
 }
 
 @media screen and (-ms-high-contrast: active),
 (-ms-high-contrast: none) {
+
     .ps {
+
         overflow: auto !important;
     }
 }
